Issue #73 refactor: 가독성을 위한 파일 분리
